Perform code cleanup and make use of modern Java features (#2442)

+++ src/main/java/mezz/jei/gui/ghost/GhostIngredientDrag.java
@@ -81,10 +81,10 @@ public class GhostIngredientDrag<T> {
 			GL11.glEnable(GL11.GL_LINE_SMOOTH);
 			GL11.glHint(GL11.GL_LINE_SMOOTH_HINT, GL11.GL_NICEST);
 			GL11.glBegin(GL11.GL_LINES);
-			float red = (float) (targetColor >> 24 & 255) / 255.0F;
-			float green = (float) (targetColor >> 16 & 255) / 255.0F;
-			float blue = (float) (targetColor >> 8 & 255) / 255.0F;
-			float alpha = (float) (targetColor & 255) / 255.0F;
+			float red = (targetColor >> 24 & 255) / 255.0F;
+			float green = (targetColor >> 16 & 255) / 255.0F;
+			float blue = (targetColor >> 8 & 255) / 255.0F;
+			float alpha = (targetColor & 255) / 255.0F;
 			RenderSystem.setShaderColor(red, green, blue, alpha);
 			GL11.glVertex3f(mouseX, mouseY, 150);
 			GL11.glVertex3f(originX, originY, 150);
